Code Tidy up
- Extracts base class for Predict and TranPredict Widgets
- Use imports rather than duplicating code
- Fixes issue with reloading labels
